Changed how render trees are constructed
Now when you add a child, you must hand in a unique_ptr so that
ownership over the type will be fully transfered.  It will give
back a shard_ptr to the base class, so accessor functions to
things like translation and rotation are still there.  But now
once you had a node as a child of another, that subtree is considered
complete.  The number of points will not change, no further children
will be added deeper in the tree, and access to any child specific
functions will be lost.

As a consequence of this, the underlying data buffers can be created
immediately upon construction of a root node object.  It's no longer
an independent call that can be forgotten or done multiple times