Changed the functions that fill in the vertex buffers to make it more unlikely to overwrite another nodes data.
In the previous implementation, each child class that overrode these
methods would have to correctly use the paremter indicating the start
index to avoid overwriting another nodes data, which would be very
error prone as more and more classes fill out this interface.  Now
a custom object giving us a view into a slice of a Vector is handed in.

Children are still responsible for not writing too much data, though
asserts in debug builds should catch this as well.