Changed how TypedRenderNode works
The implementation class now separates the type it inherits from
the type it uses when constructing a tree.  Now then implementation class
takes two variadic template packs, one for each usage.  It is only correct
to hand in the same pack twice (normal usage), or with the second pack
empty (for Pure nodes).  Two convenience aliases have been supplied so that
calling code does not need to worry about this distinction.