Added the ability to construct types
VariableFactory has a templated `temporary` function used to
create temporary variables.  The template parameter determines
what type is created, and the arguments are used as arguments
to the constructor.  Any type being created must define a static
`Create` function with a matching signature in order for this to
work.  Handing in a Variable<T> or Expression<T> will automatically
be unwrapped and call the `Create` function with a raw T.

Note, this is not going to play nicely with references or const
qualifiers.  If we need to be very precise with our language
inside opengl, this approach may have to be revisited