Separating the vao from the vbo's
The RootNode still controls the life of the VBO's, both for the actual vertex
data as well as the index buffer.  However, this data could potentially could
be used with multiple shader programs, each having a different layout for the
data.  A new RenderingTree class has been created, which is used to control
the lifecycle of vao's.  The RootNode no longer cares about compatability with
a given shader, that will be enforced by the RenderingTree