Slightly reworked how Manipulators function
Now the PureNode extends from the manipulator, so
that upon creation you have direct access to it.
This is in preparation for an upcoming change where
all children must be added at once, making it difficult
to return a manipulator after adding a node to the tree.
A new function returning a CallableReferenceWrapper
to the manipulator will be added if calling code needs
to retain control over that even after handing the
node itself over to the tree.