Looooong hiatus, not actually entire sure what is in this commit
- Much of the changes are whitespace fixes, from tab to space
- Cleaned up a few issues pointed out by clang
- Fixed bug where rotation/translation accumulation were done
  incorrectly.  The rotation of a node was erroneously being
  applied to it's own translation during accumulation

The main highlight however is the ability to set an "origin"
node via the Transform::InitManager.  When this is node is
set, the path from that node to the root is saved.  Then each
time a new rendering is requested we first walk up this path
to compute the inverse transformation, and use this as the
starting point for that particular uniform.

The benefits are two fold.  First, this means the camera can
be pegged to an object somewhere in the heirarchy for the whole
world, so even if objects move in relation to each other, the
camera can automatically track with an object. (Thing a camera
pointing at a person's face, as they sit on a merry-go-round)

Secondly, this sets up a proof of concept for how to handle
"lights".  Will still need to set up coordination between
the light and transform uniforms, but now we have a way to
translate a node's coordinate system to the global coordinate
system